owpreprocess: moved preprocess helpers into utils.
